Move AbstractJsonWebToken to Token namespace
